services/processor/requirements.txt
kafka-python==2.0.2
redis==5.2.0
prometheus-client==0.16.0
tenacity==8.2.2
python-dotenv==1.0.0
pydantic>=2.7.0,<3.0
pydantic-settings==2.11.0

# embedding + HF stack
sentence-transformers==2.6.1
huggingface-hub>=0.19.3,<1.0

# numeric stack (aligned with numpy 1.26.x)
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2

# torch pins — MUST match the API image to avoid resolver changes
torch==2.2.2
torchvision==0.17.2

# extras used by sentence-transformers
tqdm
nltk
sentencepiece
